FIX-316: Fix compilation errors after versioning implementation
🔧 COMPILATION FIXES:

Type conversions:
- messages/component.go: int32(processVersion) for ProcessMessageSubscription
- receive_task.go: int32(processVersion) for ProcessMessageSubscription
- intermediate_catch_message_handler.go: int32(processVersion) for ProcessMessageSubscription
- process_service.go: extractVersionFromKey(result.ProcessKey) for interfaces.ProcessInstanceStatus

Function deduplication:
- Removed duplicate extractVersionFromKey functions (use existing from engine.go)
- Added extractVersionFromKey to process_service.go and messages/component.go
- Cleaned unused imports (strconv, strings)

✅ Build successful: make build passes

Russian: FIX-316: Исправление ошибок компиляции после реализации версионирования

+++ src/messages/component.go
@@ -388,7 +388,7 @@ func (c *Component) handleCreateSubscription(ctx context.Context, request Messag
 		ID:                   models.GenerateID(),
 		TenantID:             payload.TenantID,
 		ProcessDefinitionKey: payload.ProcessKey,
-		ProcessVersion:       processVersion, // Use actual version from ProcessKey
+		ProcessVersion:       int32(processVersion), // Use actual version from ProcessKey
 		StartEventID:         payload.ElementID,
 		MessageName:          payload.MessageName,
 		MessageRef:           payload.MessageName,
@@ -568,3 +568,4 @@ func extractVersionFromKey(processKey string) int {
 	}
 	return 1
 }
+
